Move challenge functionality related components to challenge feature folder
