+++ frontend/src/features/challenge/components/ChallengeResponseWaitScreen.tsx
@@ -1,7 +1,7 @@
-import { TimeControl } from "../../types/gameSetup";
-import { displayTimeControl } from "../../utils/timeUtils";
-import LoadingSpinner from "../../components/common/LoadingSpinner";
-import ModalWrapper from "../../components/wrappers/ModalWrapper";
+import { TimeControl } from "../../../types/gameSetup";
+import { displayTimeControl } from "../../../utils/timeUtils";
+import LoadingSpinner from "../../../components/common/LoadingSpinner";
+import ModalWrapper from "../../../components/wrappers/ModalWrapper";
 
 import "../../styles/modals/challenge-response-wait-screen.scss";
